requirements.txt
streamlit
pytesseract
google-generativeai
PyMuPDF
Pillow
